refactor(api_server): extract family-key and envelope construction from the submit handler
Keeps handleSubmitTransactions under the nestif threshold now that the key
lookup has a fallback branch, and gives the stability rationale a place to
live next to familyPartitionKeys rather than buried in the handler. parsedItem
moves to package scope so both helpers can name it. No behavior change.
