+++ services/api_server/handlers.go
@@ -1028,11 +1028,6 @@ func (s *Server) handleSubmitTransactions(c *gin.Context) {
 	// Phase 1: parse the whole stream. Use the canonical TxID() (not a
 	// hash of the wire bytes) so Extended Format submissions key the
 	// same as the canonical txid the propagator broadcasts.
-	type parsedItem struct {
-		tx   *sdkTx.Transaction
-		raw  []byte
-		txid string
-	}
 	var parsed []parsedItem
 	offset := 0
 	for offset < len(body) {
@@ -1181,57 +1176,8 @@ func (s *Server) handleSubmitTransactions(c *gin.Context) {
 	// multi-partition topic a whole family lands on one partition and
 	// per-partition order still delivers parents before children.
 	if len(toPublish) > 0 {
-		// Keys are computed over `parsed` — the FULL submitted batch — not
-		// over toPublish. The distinction is the difference between a stable
-		// family key and a moving one.
-		//
-		// The key is min(family), so it moves whenever a member leaves the
-		// set. Members leave constantly: the Phase 3 dedup CAS drops every tx
-		// already known at a non-REJECTED status. The standard wallet pattern
-		// POST {P} then POST {P, C} therefore dedups P out, leaving the child
-		// with no in-batch parent and keying it by its own txid — onto a
-		// partition its still-in-flight parent is not on, where the
-		// dispatcher cannot hold it behind that parent. It draws
-		// TX_MISSING_PARENT and parks, where before #295's multi-partition
-		// topic it was admitted in milliseconds. The partial-produce retry
-		// path has the same shape: the members that already landed dedup out
-		// and the remainder recomputes onto a different partition.
-		//
-		// A deduped-out member can remain the family representative because
-		// the key is only a hash input; it does not have to be a txid that
-		// this request publishes.
-		allTxids := make([]string, len(parsed))
-		allInputs := make([][]string, len(parsed))
-		for i, p := range parsed {
-			allTxids[i] = p.txid
-			allInputs[i] = collectInputTXIDs(p.tx)
-		}
-		allKeys := familyPartitionKeys(allTxids, allInputs)
-		keyByTxid := make(map[string]string, len(parsed))
-		inputsByTxid := make(map[string][]string, len(parsed))
-		for i, p := range parsed {
-			if _, ok := keyByTxid[p.txid]; ok {
-				continue // duplicate txid in one body: first occurrence wins
-			}
-			keyByTxid[p.txid] = allKeys[i]
-			inputsByTxid[p.txid] = allInputs[i]
-		}
-
-		msgs := make([]kafka.KeyValue, 0, len(toPublish))
-		for _, p := range toPublish {
-			key, ok := keyByTxid[p.txid]
-			if !ok {
-				key = p.txid
-			}
-			msgs = append(msgs, kafka.KeyValue{
-				Key: key,
-				Value: map[string]interface{}{
-					"txid":        p.txid,
-					"raw_tx":      p.raw,
-					"input_txids": inputsByTxid[p.txid],
-				},
-			})
-		}
+		keyByTxid, inputsByTxid := familyKeysBySubmittedTxid(parsed)
+		msgs := propagationMessages(toPublish, keyByTxid, inputsByTxid)
 		if err := s.producer.SendBatch(ctx, kafka.TopicPropagation, msgs); err != nil {
 			if errors.Is(err, kafka.ErrBrokerBackpressure) {
 				s.logger.Warn("batch submit rejected: kafka backpressure", zap.Int("count", len(msgs)))
